+++ src/ui/panes/albums.rs
@@ -47,7 +47,9 @@ impl AlbumsPane {
 
 impl Pane for AlbumsPane {
     fn render(&mut self, frame: &mut Frame, area: Rect, _context: &AppContext) -> Result<()> {
-        self.browser.render(area, frame.buffer_mut(), &mut self.stack);
+        self.browser
+            .set_filter_input_active(self.filter_input_mode)
+            .render(area, frame.buffer_mut(), &mut self.stack);
 
         Ok(())
     }