+++ src/ui/panes/playlists.rs
@@ -51,7 +51,9 @@ impl PlaylistsPane {
 
 impl Pane for PlaylistsPane {
     fn render(&mut self, frame: &mut Frame, area: Rect, _context: &AppContext) -> Result<()> {
-        self.browser.render(area, frame.buffer_mut(), &mut self.stack);
+        self.browser
+            .set_filter_input_active(self.filter_input_mode)
+            .render(area, frame.buffer_mut(), &mut self.stack);
 
         Ok(())
     }